Replace integration tests with unit tests using faketime from env to reduce faketime dependency
